### 2025-09-25 #9080 @editorialbot
[PRE REVIEW]: A Python API for OTTER
**Submitting author:** @noahfranz13 (<a href="http://orcid.org/0000-0003-4537-3575">Noah Franz</a>)
**Repository:** https://github.com/astro-otter/otter
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.4.1
