### 2025-09-26 #9080 @warrickball
Hi @noahfranz13, thanks for your submission! All the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review for now. I noticed, however, that this is a joint submission with AAS Journals, so let us know here how the AAS Journals review is progressing.

The paper is failing to compile because you have tabs in the YAML metadata block. If you fix those, you can try again to create the paper draft by commenting in this issue with `@editorialbot generate pdf`.

Also note that the paper is pushing our length limits a bit. [Our docs say](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain) papers "should be between 250-1000 words". That's not a hard upper limit but it would be good to cut 100-200 words if possible. You can comment here with `@editorialbot check repository` to run the checks that include the word count.
