### 2025-09-27 424c3b8
first try at a query for photometry wtihout upper limit
