### 2025-09-27 d1b1c5a
make the JOSS paper more concise and remove tabs in the yaml header
