+++ paper/paper.md
@@ -6,10 +6,10 @@ authors:
     affiliation: 1
   - name: Kate D Alexander
     orcid: 0000-0002-8297-2473
-	affiliation: 1
+    affiliation: 1
   - name: Sebastian Gomez
     orcid: 0000-0001-6395-6702
-	affiliation: 2
+    affiliation: 2
 affiliations:
   - name: Department of Astronomy and Steward Observatory, University of Arizona, 933 North Cherry Avenue, Tucson, AZ 85721-0065, USA
     index: 1
@@ -27,7 +27,7 @@ aas-journal: Astrophysical Journal
 The Open mulTiwavelength Transient Event Repository (or "OTTER") is a new catalog of
 published transient data. Here we present a thick Python wrapper on the
 REpresentational State Transfer (REST) application programming interface (API) built-in to the OTTER backend database
-(hereafter the "OTTER API"). Since the OTTER backend is built on the document database
+(the "OTTER API"). Since the OTTER backend is built on the document database
 ArangoDB, using the REST API directly requires learning the Arango Query Language
 (AQL). Since AQL has a niche user base, OTTER users unfamiliar with it may face a roadblock to programmatic access of
 OTTER. To overcome this barrier, we created the OTTER Python API to make programmatic access easy
@@ -82,20 +82,18 @@ data. However, the API endpoints expect queries in the syntax of the "Arango Que
 Language" (AQL). Learning a new query language creates a barrier for user programmatic
 access to the indispensable dataset available in the OTTER catalog.
 
-To help overcome this barrier, we present an object-oriented Python API for access to
+To help overcome this barrier, we present a Python API for access to
 the OTTER dataset. This API acts as a thick wrapper on the AQL-based API, with many additions
-that make it easier to access and analyze the dataset. Some of the more notable
-components include the following:
+that make it easier to access and analyze the dataset. Some of these features include:
 
 * In OTTER we store photometry as close to the actual
-  published value as possible. This makes the data more reproducible since it does
-  not rely on complex processing scripts that may have unknown bugs. However, this also
+  published value as possible to make the data more reproducible. However, this also
   means that the data is not stored in consistent units (but the unit of the photometry
   point is stored). In the OTTER API we automatically convert the photometry into the user-requested units.
   Specifically, the conversion is done in the
   `Otter.get_phot` and `Transient.clean_photometry` methods which use
-  astropy [@astropy_collaboration_astropy_2013; @astropy_collaboration_astropy_2018;
-  @astropy_collaboration_astropy_2022] and synphot [@stsci_development_team_synphot_2018].
+  `astropy` [@astropy_collaboration_astropy_2013; @astropy_collaboration_astropy_2018;
+  @astropy_collaboration_astropy_2022] and `synphot` [@stsci_development_team_synphot_2018].
 * The same raw data from an astronomical observation may be reduced[^2] by
   multiple, distinct, teams. Depending on the differences in the reduction methodology this may
   produce different flux measurements. If this is the case, we store both flux
@@ -106,10 +104,7 @@ components include the following:
 * Sometimes users want to quickly view the photometry for a specific transient
   event as either a light curve (flux as a function of time) or a spectral energy distribution (flux
   as a function of wavelength, frequency, or energy; i.e., an "SED"). In the `plotter` module of the OTTER API we
-  provide numerous methods for quickly and automatically plotting the photometry. The
-  most powerful of these methods (albeit, the least flexible) is the `query_quick_view`
-  method which allows the user to query the database for e.g., the name of a specific transient,
-  and then returns both a light curve and SED of the associated photometry [@hunter_matplotlib_2007; @plotly].
+  provide numerous methods for quickly and automatically plotting the photometry [@hunter_matplotlib_2007; @plotly].
[…]
