### 2025-09-27 #9080 @noahfranz13
Thanks @warrickball!

> [Our docs say](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain) papers "should be between 250-1000 words". That's not a hard upper limit but it would be good to cut 100-200 words if possible.

Sorry about that, I read the docs so many times and I must have missed that! I've cut ~180 words out and gotten the paper length down to <1500 words. I hope that's okay!

> The paper is failing to compile

I'm also sorry about this! I was consistently able to compile it locally with docker and in github using `inara:latest`. I untabified it and it seems to compile now though!
