### 2025-09-28 #9080 @noahfranz13
I know this is currently waitlisted, but one reviewer suggestion from the [list of signed up reviewers](https://reviewers.joss.theoj.org/reviewers/search) is Heloise Stevance (Github: HeloiseS). I've emailed with them and know they are working on similar projects and will likely give a constructive review.
