### 2025-09-30 c079107
fix the journal names in the bibtex
