+++ paper/paper.bib
@@ -279,21 +279,23 @@ ADS Bibcode: 2018ApJ...859L..20D},
 	file = {arXiv Fulltext PDF:/home/nfranz/snap/zotero-snap/common/Zotero/storage/KXXFEQPX/Tanoglidis et al. - 2023 - Transformers for scientific data a pedagogical re.pdf:application/pdf;arXiv.org Snapshot:/home/nfranz/snap/zotero-snap/common/Zotero/storage/T9D63D46/2310.html:text/html},
 }
 
-@misc{stein_texttttdescore_2023,
-	title = {\${\textbackslash}texttt\{tdescore\}\$: {An} {Accurate} {Photometric} {Classifier} for {Tidal} {Disruption} {Events}},
-	shorttitle = {\${\textbackslash}texttt\{tdescore\}\$},
-	url = {http://arxiv.org/abs/2312.00139},
-	doi = {10.48550/arXiv.2312.00139},
-	abstract = {Optical surveys have become increasingly adept at identifying candidate Tidal Disruption Events (TDEs) in large numbers, but classifying these generally requires extensive spectroscopic resources. We here present \${\textbackslash}texttt\{tdescore\}\$, a simple photometric classifier that is trained using a systematic census of \${\textbackslash}sim\$3000 nuclear transients from the Zwicky Transient Facility (ZTF). The sample is highly imbalanced, with TDEs representing \${\textless}\$2\% of the total. \${\textbackslash}texttt\{tdescore\}\$ is nonetheless able to reject non-TDEs with 99.6\% accuracy, yielding a sample of probable TDEs with completeness of 77.0\% and a purity of 80.3\%. \${\textbackslash}texttt\{tdescore\}\$ is thus substantially better than any available TDE photometric classifier scheme in the literature, and performs comparably well to the single-epoch spectroscopy as a method for classifying ZTF nuclear transients, despite relying solely on ZTF data and multi-wavelength catalogue crossmatching. In a novel extension, we use 'SHapley Additive exPlanations' (SHAP) to provide a human-readable justification for each individual \${\textbackslash}texttt\{tdescore\}\$ classification, enabling users to understand and form opinions about the underlying classifier reasoning. \${\textbackslash}texttt\{tdescore\}\$ serves as a model for photometric identification of TDEs with time-domain surveys, such as the upcoming Rubin observatory.},
-	urldate = {2023-12-04},
-	publisher = {arXiv},
-	author = {Stein, Robert and Mahabal, Ashish and Reusch, Simeon and Graham, Matthew and Kasliwal, Mansi M. and Kowalski, Marek and Gezari, Suvi and Hammerstein, Erica and Nakoneczny, Szymon J. and Nicholl, Matt and Sollerman, Jesper and van Velzen, Sjoert and Yao, Yuhan and Laher, Russ R. and Rusholme, Ben},
-	month = nov,
-	year = {2023},
-	note = {arXiv:2312.00139 [astro-ph]},
-	keywords = {notion, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Instrumentation and Methods for Astrophysics},
-	annote = {Comment: 12 pages, 5 figures, 1 table},
[…]
