### 2025-09-30 #9080 @noahfranz13
> note that I think some of your references still have NASA ADS macros in the journal field (e.g. \apj for The Astrophysical Journal)

This is fixed in that newest pdf.
