### 2025-10-08 8c26a95
update pydantic version requirement and reversion
