### 2025-10-09 31dbc7e
add query parameter that can be used to check if the transient has a detection
