### 2025-10-09 1239819
add wave_det keyword for searching for detections in specific wavelength regimes
