### 2025-11-11 2ac26f9
replace nans with nulls in xray model parameters in the csvs
