### 2025-11-11 43c946f
update tests to use deployment URL
