### 2025-11-19 #9080 @warrickball
Hi @jwuphysics and @cmbiwer! Would you be able to review this submission?

I presume you're familiar with the process because you've reviewed for JOSS before but let me know if you have any questions about our open, [checklist-driven](https://joss.readthedocs.io/en/latest/review_checklist.html) and preferably iterative review process, let me know.
