### 2025-11-26 #9516 @editorialbot
[REVIEW]: A Python API for OTTER
**Submitting author:** @noahfranz13 (<a href="http://orcid.org/0000-0003-4537-3575">Noah Franz</a>)
**Repository:** https://github.com/astro-otter/otter
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.0
