### 2025-12-03 81189ca
add photometry table column descriptions (Issues #26 and #31
