### 2025-12-08 a57e2c2
remove negative flux filter from clean_photometry per guidance from Sjoert over email
