+++ src/otter/_version.py
@@ -2,4 +2,4 @@
 Just define the package version in one place
 """
 
-__version__ = "0.4.2"
+__version__ = "0.5.0"
