### 2025-12-09 1db4bff
update some plotting issues created by the last commit (this is why we have tests)
