### 2025-12-09 271c235
update test suite to handle slow external server connections cleaner
