### 2025-12-09 eda2177
make pytest mark windows and macos safe
