### 2025-12-09 6069b06
add contribution and support guidelines (openjournals/joss-reviews#9516)
