+++ docs/contributing.rst
@@ -1,0 +1,23 @@
+Seeking Support, Reporting Issues, and Contributing
+---------------------------------------------------
+
+Reporting Issues, Providing Feedback, and Seeking Support
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+Please report any bugs, issues, or suggestions by opening an issue at
+https://github.com/astro-otter/otter/issues/new/choose. You should also feel
+free to open an issue just with questions if something in the documentation
+is unclear. Typically, questions like these are very useful and will lead to
+additions or clarifications on our documentation page! We monitor new
+issues opened here and will provide additional guidance and/or next steps.
+
+Contributing
+^^^^^^^^^^^^
+Please use the following Git/GitHub workflow if you wish to make contributions:
+
+#. Fork this repository
+#. Following the Developer Installation above to install the API locally. Feel free to open an issue or reach out with any problems.
+#. When you have finished your contribution (or have reached a stable point in your development), please open a Pull Request from your fork.
+
+We monitor the opened pull requests and will discuss any implementation details
+with you there, if necessary.
