### 2025-12-09 c61c157
fix paper typos and issues mentioned in openjournals/joss-reviews#9516
