### 2025-12-09 b1f5058
add more details about appropriate units in get_phot and clean_photometry (#33)
