### 2025-12-09 81784a1
add details on the class_confidence_threshold argument for Otter.query #33
