+++ src/otter/io/otter.py
@@ -352,8 +352,23 @@ class Otter(Database):
                                 published classification in OTTER. If classifications
                                 disagree for a transient, it will be filtered out.
             classification (str): A classification string to search for
-            class_confidence_threshold (float): classification confidence cutoff for
-                                                query, between 0 and 1. Default is 0.
+            class_confidence_threshold (float): classification confidence cutoff.
+                                                Default is 0. Any classifications with a
+                                                confiedence less than
+                                                this value will be ignored. In general,
+                                                (1) C = 0 means this classification is
+                                                unverified, (2) C = 1 means this
+                                                classification is based only on
+                                                photometry, (3) C = 2 means this
+                                                classification is based on a TNS (not
+                                                peer-reviewed) spectrum, (4) C >= 3
+                                                means this classification is based on
+                                                at least one peer-reviewed optical
+                                                spectrum. Setting this flag to 3 should
+                                                produce the most pure sample, but might
+                                                miss some events that are very
+                                                confidently classified based on
+                                                photometry (e.g., X-ray observations).
             has_det (bool): This can be set to true to only search for transients that
                             have a detection in their photometry. It can be used in
                             conjunction with e.g., `has_radio_phot=True` to search for
