### 2025-12-09 ef1d198
add details on the class_confidence_threshold argument for Otter.query
