### 2025-12-12 c40287e
Test: Attempt direct push to main
