+++ src/otter/README.md
@@ -1,1 +1,0 @@
-Test change Fri Dec 12 08:54:31 AM MST 2025
