### 2025-12-12 cafbe26
fix: add new semver.yml
