+++ .github/workflows/semver.yml
@@ -1,25 +1,62 @@
-name: sem-ver
+name: Continuous Delivery
+
 on:
   push:
     branches:
       - main
 
+# default: least privileged permissions across all jobs
+permissions:
+  contents: read
+
 jobs:
-  main:
+  release:
     runs-on: ubuntu-latest
+    concurrency:
+      group: ${{ github.workflow }}-release-${{ github.ref_name }}
+      cancel-in-progress: false
+
     permissions:
-      contents: write # Needed for creating tags and pushing changes
+      contents: write
 
     steps:
-      - name: Checkout
-        uses: actions/checkout@v5
+      # Note: We checkout the repository at the branch that triggered the workflow.
+      # Python Semantic Release will automatically convert shallow clones to full clones
+      # if needed to ensure proper history evaluation. However, we forcefully reset the
+      # branch to the workflow sha because it is possible that the branch was updated
+      # while the workflow was running, which prevents accidentally releasing un-evaluated
+      # changes.
+      - name: Setup | Checkout Repository on Release Branch
+        uses: actions/checkout@v4
         with:
-          token: ${{ secrets.SEMVER_PAT_TOKEN }}
-          fetch-depth: 0
+          ref: ${{ github.ref_name }}
 
-      - name: Semantic Release
-        uses: cycjimmy/semantic-release-action@v6
-        env:
-          # 4. Pass the PAT to the GITHUB_TOKEN environment variable
-          # (Ensure the secret name is correct: SEMVER_PAT_TOKEN)
-          GITHUB_TOKEN: ${{ secrets.SEMVER_PAT_TOKEN }}
+      - name: Setup | Force release branch to be at workflow sha
+        run: |
+          git reset --hard ${{ github.sha }}
+
+      - name: Action | Semantic Version Release
+        id: release
+        # Adjust tag with desired version if applicable.
+        uses: python-semantic-release/python-semantic-release@v10.5.2
+        with:
+          github_token: ${{ secrets.GITHUB_TOKEN }}
+          git_committer_name: "github-actions"
+          git_committer_email: "actions@users.noreply.github.com"
+
+      - name: Publish | Upload to GitHub Release Assets
+        uses: python-semantic-release/publish-action@v10.5.2
+        if: steps.release.outputs.released == 'true'
+        with:
+          github_token: ${{ secrets.GITHUB_TOKEN }}
+          tag: ${{ steps.release.outputs.tag }}
+
+      - name: Upload | Distribution Artifacts
+        uses: actions/upload-artifact@v4
+        with:
+          name: distribution-artifacts
+          path: dist
+          if-no-files-found: error
+
+    outputs:
+      released: ${{ steps.release.outputs.released || 'false' }}
