### 2025-12-12 85deeff
fix: try using a PAT instead of GITHUB_TOKEN
