### 2025-12-12 3a02c1f
fix: following python-semantic-version docs add the token argument to actions/checkout
