+++ .github/workflows/semver.yml
@@ -30,6 +30,7 @@ jobs:
         uses: actions/checkout@v6
         with:
           ref: ${{ github.ref_name }}
+          token: ${{ secrets.SEMVAR_PAT_TOKEN }}
           fetch-depth: 0
 
       - name: Setup | Force release branch to be at workflow sha
