### 2025-12-12 00bb113
fix: allow for versions to start with 0 with semantic-versioning
