+++ pyproject.toml
@@ -127,11 +127,11 @@ commit_message = "{version}\n\nAutomatically generated by python-semantic-releas
 commit_parser = "conventional"
 logging_use_named_masks = false
 major_on_zero = true
-allow_zero_version = false
+allow_zero_version = true
 no_git_verify = false
 tag_format = "v{version}"
 add_partial_tags = false
-version_variable = "src/otter/_version.py:__version__"
+version_variables = ["src/otter/_version.py:__version__"]
 
 [tool.semantic_release.branches.main]
 match = "main"
