### 2025-12-12 4666eaa
fix: a typo in the semver workflow
