### 2025-12-12 7ec1d6c
0.6.0
Automatically generated by python-semantic-release
