+++ CHANGELOG.md
@@ -1,3 +1,42 @@
 # CHANGELOG
 
 <!-- version list -->
+
+## v0.6.0 (2025-12-12)
+
+### Bug Fixes
+
+- A typo in the semver workflow
+  ([`4666eaa`](https://github.com/astro-otter/otter/commit/4666eaa86fa4c9524daa0954c9f5df58301e0ad7))
+
+- Add new semver.yml
+  ([`cafbe26`](https://github.com/astro-otter/otter/commit/cafbe2648181ab067b4ed2976fd3931370d36033))
+
+- Allow for versions to start with 0 with semantic-versioning
+  ([`00bb113`](https://github.com/astro-otter/otter/commit/00bb113133bf358202b83e3e9f6ef04ee8bf5418))
+
+- Following python-semantic-version docs add the token argument to actions/checkout
+  ([`3a02c1f`](https://github.com/astro-otter/otter/commit/3a02c1f36e587f5134f827f4788a9e0b09815c68))
+
+- Try adding in the repository name explicitly
+  ([`43deea5`](https://github.com/astro-otter/otter/commit/43deea50cab952d2387c21f6b013a52059219c20))
+
+- Try again with the semver workflow
+  ([`6b7f948`](https://github.com/astro-otter/otter/commit/6b7f94866eb58a80f2fd17a6db5d3de6893beeb8))
+
+- Try again with the semver workflow
+  ([`c51e999`](https://github.com/astro-otter/otter/commit/c51e99971d768524b5d7786e05c1ecfec2c2c9f8))
+
+- Try to fix the semver workflow file
+  ([`9bb8bce`](https://github.com/astro-otter/otter/commit/9bb8bce1c642529f4c4658a14f1281d466b24dee))
+
+- Try using a PAT instead of GITHUB_TOKEN
+  ([`85deeff`](https://github.com/astro-otter/otter/commit/85deeffcd5fe638c865eab45cdc3f034cc1d177e))
+
+- Update to use more recent versions of github actions
+  ([`84926d8`](https://github.com/astro-otter/otter/commit/84926d87c5822b9a1163fb4fcb4fb58e06067b94))
+
+### Features
+
+- Try to add semantic versioning automatically
+  ([`33f7664`](https://github.com/astro-otter/otter/commit/33f7664930b3742753fb67ad1906085f38515653))
