### 2025-12-12 ec32d65
fix: ensure we don't run the semver workflow in an infinite loop
