+++ .github/workflows/semver.yml
@@ -54,12 +54,5 @@ jobs:
           github_token: ${{ secrets.SEMVER_PAT_TOKEN }}
           tag: ${{ steps.release.outputs.tag }}
 
-      - name: Upload | Distribution Artifacts
-        uses: actions/upload-artifact@v5
-        with:
-          name: distribution-artifacts
-          path: dist
-          if-no-files-found: error
-
     outputs:
       released: ${{ steps.release.outputs.released || 'false' }}
