+++ pyproject.toml
@@ -123,7 +123,7 @@ quote-style = "double"
 [tool.semantic_release]
 assets = []
 build_command_env = []
-commit_message = "{version}\n\nAutomatically generated by python-semantic-release"
+commit_message = "v{version}\n\nAutomatically generated by python-semantic-release [skip actions]"
 commit_parser = "conventional"
 logging_use_named_masks = false
 major_on_zero = true
