### 2025-12-12 649491a
v0.6.1
Automatically generated by python-semantic-release [skip actions]
