+++ src/otter/_version.py
@@ -2,4 +2,4 @@
 Just define the package version in one place
 """
 
-__version__ = "0.6.0"
+__version__ = "0.6.1"
