### 2025-12-12 3b263ab
fix: add specific paths to only run CI unit tests when those files are edited
