+++ .github/workflows/main.yml
@@ -3,6 +3,9 @@ name: CI
 on:
   pull_request:
   push:
+    paths:
+      - src/*
+      - tests/*
     branches:
       - main
 
