### 2025-12-12 68afdf2
v0.6.2
Automatically generated by python-semantic-release [skip actions]
