### 2025-12-12 d798bbd
feat: standardize UVOIR filter names (Issue #21)
