### 2025-12-12 341fd2c
feat: add automatic MW dust extinction correction to Transient.clean_photometry
