### 2025-12-12 c6e5e33
fix: minor bug in photometry de-reddening
