### 2025-12-12 63b7d0d
fix: fully address issue #32
