### 2025-12-12 7fba7a6
fix: redshift is now cast as a float (Issue #20)
