### 2025-12-12 ce89ccc
fix: Allow users to pass numpy arrays of names to Otter.query/Otter.get_meta (Issue #19)
