### 2025-12-16 8e8c268
update to use local dustmap files for MW extinction corrections
