### 2025-12-16 0278eab
update the init script to work with dustmaps.sfd
