### 2025-12-16 07149b3
add updated installation instructions for windows users
