### 2025-12-16 a365f65
remove windows unit tests since we no longer natively support windows
