+++ .github/workflows/main.yml
@@ -19,7 +19,6 @@ jobs:
           - '3.11'
         os:
           - ubuntu-latest
-          - windows-latest
           - macos-latest
 
     runs-on: ${{ matrix.os }}
