### 2025-12-16 3422ed3
v0.7.0
Automatically generated by python-semantic-release [skip actions]
