### 2025-12-17 a1e918f
add announcement banner to docs page with newest version of the API
