### 2026-01-05 #9516 @noahfranz13
> Could you reduce the paper.bib file to the entries actually used in the JOSS paper? Thanks!

This should be fixed in the newest commit to the paper branch.

> Make a new tagged release of your software, and list the version tag of the archived version here.

I made the tag/release v1.0.0 that includes all of these changes.

> Please list the DOI of the archive here.

Here is the Zenodo DOI: https://doi.org/10.5281/zenodo.18155961. I updated the title and author list to match the JOSS submission. The software is licensed under the MIT license for both the Zenodo and github repositry.

Let me know if you need anything else from me @warrickball!
