### 2026-01-05 4f8f4ff
add back dois to the bib file
