+++ paper/paper.bib
@@ -11,8 +11,7 @@
     {Multi-Component} {Outflows} vs. {Host} {Contamination}}},
   journal = {arXiv e-prints},
   pages = {arXiv:2509.14317},
-  year = {2025},
-  month = {sep},
+  date = {2025-09},
   url = {https://arxiv.org/abs/2509.14317}
 }
 @article{franz_otter_2025,
@@ -26,8 +25,7 @@
     {Disruption} {Event} {Catalog}}},
   journal = {arXiv e-prints},
   pages = {arXiv:2509.05405},
-  year = {2025},
-  month = {sep},
+  date = {2025-09},
   url = {https://arxiv.org/abs/2509.05405}
 }
 @article{auchettl_new_2017,
@@ -40,9 +38,36 @@
   volume = {838},
   number = {2},
   pages = {149},
-  year = {2017},
-  month = {apr},
-  url = {https://iopscience.iop.org/article/10.3847/1538-4357/aa633b}
+  date = {2017-04},
+  urldate = {2024-07-26},
+  url = {https://iopscience.iop.org/article/10.3847/1538-4357/aa633b},
+  doi = {10.3847/1538-4357/aa633b},
+  issn = {0004-637X, 1538-4357},
+  abstract = {We perform a comprehensive study of the X-ray emission
+    from 70 transient sources that have been classiﬁed as tidal
+    disruption events (TDEs) in the literature. We explore the
+    properties of these candidates, using nearly three decades of X-ray
+    observations to quantify their properties and characteristics. We
+    ﬁnd that the emission from X-ray TDEs increase by two to three
+    orders of magnitude, compared to pre-ﬂare constraints. These
+    emissions evolve signiﬁcantly with time, and decay with power-law
+    indices that are typically shallower than the canonical t−5/3 decay
+    law, implying that X-ray TDEs are viscously delayed. These events
+    exhibit enhanced (relative to galactic) column densities and are
+    quite soft in nature, with no strong correlation between the amount
+    of detected soft and hard emission. At their peak, jetted events
+    have an X-ray to optical ratio ?1, whereas non-jetted events have a
+    ratio ∼1, which suggests that these events undergo reprocessing at
+    different rates. X-ray TDEs have long T90 values, consistent with
+    what would be expected from a viscously driven accretion disk formed
+    by the disruption of a main-sequence star by a black hole with a
+    mass {\textless{}}107 Me. The isotropic luminosities of X-ray TDEs
+    are bimodal, such that jetted and non-jetted events are separated by
+    a “reprocessing valley” that we suggest is naturally populated by
+    optical/UV TDEs that most likely produce X-rays, but this emission
+    is “veiled” from observations due to reprocessing. Our results
+    suggest that non-jetted X-ray TDEs likely originate from partial
+    disruptions and/or disruptions of low-mass stars.}
 }
 @article{guolo_systematic_2024,
   author = {Guolo, Muryel and Gezari, Suvi and Yao, Yuhan and Van
@@ -56,9 +81,36 @@
   volume = {966},
   number = {2},
   pages = {160},
-  year = {2024},
-  month = {may},
-  url = {https://iopscience.iop.org/article/10.3847/1538-4357/ad2f9f}
+  date = {2024-05},
+  urldate = {2024-07-23},
+  url = {https://iopscience.iop.org/article/10.3847/1538-4357/ad2f9f},
+  doi = {10.3847/1538-4357/ad2f9f},
+  issn = {0004-637X, 1538-4357},
+  abstract = {AbstractWe present a systematic analysis of the X-ray
+    emission of a sample of 17 optically selected, X-ray-detected tidal
+    disruption events (TDEs) discovered between 2014 and 2021. The X-ray
[…]
