### 2026-01-05 #9516 @noahfranz13
Sorry about that @warrickball! It looks like they should be fixed now! I think the ones under the "SKIP DOIs" header are all arXiv pre-prints which I'm guessing is why they don't have a doi? But let me know if I'm missing something.
