### 2026-01-07 2e99290
Initial editorial changes for publication in JOSS
