+++ paper/paper.bib
@@ -6,13 +6,14 @@
     Miller-Jones, James C. A. and Krips, Melanie and Ramirez-Ruiz,
     Enrico and Sand, David J. and Saxton, Richard and Shrestha, Manisha
     and van Velzen, Sjoert},
-  title = {{{Dichotomy} in {Long-Lived} {Radio} {Emission} from {Tidal}
-    {Disruption} {Events} {AT} 2020zso and {AT} 2021sdu:
-    {Multi-Component} {Outflows} vs. {Host} {Contamination}}},
+  title = {Dichotomy in Long-Lived Radio Emission from Tidal
+    Disruption Events {AT 2020zso} and {AT 2021sdu}:
+    Multi-Component Outflows vs.~{host} Contamination},
   journal = {arXiv e-prints},
   pages = {arXiv:2509.14317},
   date = {2025-09},
-  url = {https://arxiv.org/abs/2509.14317}
+  url = {https://arxiv.org/abs/2509.14317},
+  doi = {10.48550/arXiv.2509.14317}
 }
 @article{franz_otter_2025,
   author = {Franz, Noah and Alexander, Kate D and Gomez, Sebastian and
@@ -20,20 +21,21 @@
     Earl, Nicholas and Gezari, Suvi and Karmen, Mitchell and Margutti,
     Raffaella and Pearson, Jeniveve and Villar, V. Ashley and Zabludoff,
     Ann I},
-  title = {{{The} {Open} {mulTiwavelength} {Transient} {Event}
-    {Repository} {(OTTER):} {Infrastructure} {Release} and {Tidal}
-    {Disruption} {Event} {Catalog}}},
+  title = {The {Open} {mulTiwavelength} {Transient} {Event}
+    {Repository} {(OTTER)}: Infrastructure Release and Tidal
+    Disruption Event Catalog},
   journal = {arXiv e-prints},
   pages = {arXiv:2509.05405},
   date = {2025-09},
-  url = {https://arxiv.org/abs/2509.05405}
+  url = {https://arxiv.org/abs/2509.05405},
+  doi = {10.48550/arXiv.2509.05405}
 }
 @article{auchettl_new_2017,
   author = {Auchettl, Katie and Guillochon, James and Ramirez-Ruiz,
     Enrico},
-  title = {New {{Physical}} {{Insights}} about {{Tidal}} {{Disruption}}
-    {{Events}} from a {{Comprehensive}} {{Observational}} {{Inventory}}
-    at {{X}}-{{Ray}} {{Wavelengths}}},
+  title = {New Physical Insights about Tidal Disruption
+    Events from a Comprehensive Observational Inventory
+    at {X}-Ray Wavelengths},
   journal = {The Astrophysical Journal},
   volume = {838},
   number = {2},
@@ -73,10 +75,10 @@
   author = {Guolo, Muryel and Gezari, Suvi and Yao, Yuhan and Van
     Velzen, Sjoert and Hammerstein, Erica and Cenko, S. Bradley and
     Tokayer, Yarone M.},
-  title = {A {{Systematic}} {{Analysis}} of the {{X}}-{{Ray}}
-    {{Emission}} in {{Optically}} {{Selected}} {{Tidal}} {{Disruption}}
-    {{Events}}: {{Observational}} {{Evidence}} for the {{Unification}}
-    of the {{Optically}} and {{X}}-{{Ray}}-Selected {{Populations}}},
+  title = {A Systematic Analysis of the {X}-Ray
+    Emission in Optically Selected Tidal Disruption
+    Events: Observational Evidence for the Unification
+    of the Optically and {X}-Ray-Selected Populations},
   journal = {The Astrophysical Journal},
   volume = {966},
[…]
