### 2026-01-08 #9516 @warrickball
Thanks for the quick merge! The metadata looks good, so now we just need the AAS DOIs. Once they're available and added to the paper source, I'll do another test run and, if it still looks good, publish the paper. Almost there!
