### 2026-01-19 6066926
fix: bug in filter name cleanup
