+++ src/otter/io/transient.py
@@ -472,6 +472,7 @@ class Transient(MutableMapping):
         obs_type: str = None,
         deduplicate: Callable | None = None,
         correct_for_mw_dust: bool = True,
+        cleanup_uvoir_filternames: bool = True,
     ) -> pd.DataFrame:
         """
         Ensure the photometry associated with this transient is all in the same
@@ -510,6 +511,9 @@ class Transient(MutableMapping):
                                         and the Gordon+23 extinction curve assuming
                                         R_V=3.1. Note that this will only correct
                                         photometry in the range 0.0912 - 32 um!
+            cleanup_uvoir_filternames (bool): If True (the default) we will try to
+                                              cleanup the UV/Optical/IR filter names to
+                                              make them more standard.
         Returns:
             A pandas DataFrame of the cleaned up photometry in the requested units
         """
@@ -891,10 +895,26 @@ class Transient(MutableMapping):
 
         outdata["upperlimit"] = outdata.apply(is_upperlimit, axis=1)
 
-        # clean up filter names
-        outdata.loc[outdata.obs_type == "uvoir", "filter_name"] = outdata.loc[
-            outdata.obs_type == "uvoir", "filter_name"
-        ].apply(self._standardize_filter_names)
+        # clean up the UV/Optical/IR filter names
+        if "uvoir" in outdata.obs_type.unique() and cleanup_uvoir_filternames:
+            mask = outdata.obs_type == "uvoir"
+            new_filter_names_telescopes = outdata.loc[mask, "filter_name"].apply(
+                self._standardize_filter_names
+            )
+
+            _temp_tele_col = "_temp_new_telescope_name"
+            new_filter_names, telescopes = list(
+                zip(*new_filter_names_telescopes.tolist())
+            )
+            outdata.loc[mask, "filter_name"] = new_filter_names
+            outdata.loc[mask, _temp_tele_col] = telescopes
+            outdata.loc[mask, "telescope"] = outdata.loc[mask].apply(
+                lambda row: (
+                    row[_temp_tele_col] if pd.isna(row.telescope) else row.telescope
+                ),
+                axis=1,
+            )
+            outdata.drop(labels=_temp_tele_col, inplace=True, axis=1)
 
         # perform some more complex deduplication of the dataset
         if deduplicate:
@@ -1004,7 +1024,7 @@ class Transient(MutableMapping):
     ) -> list[str]:
         """
         This (private) method is used to clean up the filter names. As an example,
-        we want "r.ztf" = "r.ZTF" = "r" but NOT EQUAL to "R" (since capital
+        we want "r.ztf" = "r.ZTF" = "ztf.r" = "r" but NOT EQUAL to "R" (since capital
         filter names mean something different!). But here's another fun one,
         we want "UVM2.uvot" = "uvm2.uvot" = "uvm2.UVOT" = "UVM2". That one is tricky
         since the sdss and johnson-counsins filters *are* different capitalizations
@@ -1020,16 +1040,33 @@ class Transient(MutableMapping):
             "w4",
         }  # these lowercase filters should be converted to upper case
 
+        _telescope_names = {
+            "ztf",
+            "ps",
+        }  # these are some telescope names that may be in the filter names
+
         newfilt = filt
+        tele = None
         for delim in delimiters:
-            newfilt = newfilt.split(delim)[0]
+            newfilt_test = newfilt.split(delim)
+            if len(newfilt_test) != 2:
+                continue  # this delimiter was not found, so we don't do anything
+
+            newfilt_option1, newfilt_option2 = newfilt_test
+            if newfilt_option1.lower() in _telescope_names:
+                newfilt = newfilt_option2
+                tele = newfilt_option1
+            else:
[…]
