### 2026-01-19 ea1fe31
fix: add argument to clean photometry that allows users to drop not host subtracted data
