### 2026-01-19 3ef876c
fix: allow users to pass additional kwargs to Transient.clean_photometry from Otter.get_phot
