### 2026-01-19 0227fb0
fix: a bug in the photometry deduplication code
