### 2026-01-19 83ec951
fix: this is a patch for when date_min and date_max are null rather than NaN
