### 2026-01-19 dfc7229
fix: ensure the date_min/date_max cols are in the photometry df before filtering on them
