+++ src/otter/io/transient.py
@@ -1157,8 +1157,9 @@ class Transient(MutableMapping):
             # fall inside the same range
             grp["_mean_dates"] = grp.apply(cls._convert_dates, axis=1)
 
-            grp.date_min = grp.date_min.replace("null", np.nan)
-            grp.date_max = grp.date_max.replace("null", np.nan)
+            if "date_min" in grp:
+                grp.date_min = grp.date_min.replace("null", np.nan)
+                grp.date_max = grp.date_max.replace("null", np.nan)
 
             if "date_min" in grp and not np.all(pd.isna(grp.date_min)):
                 # then do the conversion
