### 2026-01-19 26cf4e7
v1.0.1
Automatically generated by python-semantic-release [skip actions]
