### 2026-01-19 663d121
v1.0.2
Automatically generated by python-semantic-release [skip actions]
