### 2026-01-23 e51c254
[feat] add wiserep queries to the OTTER API
