### 2026-01-23 2036c21
feat: acatch an re-raise a better error if the object is not on wiserep
