### 2026-01-23 8eb2ccc
fix:adad a testto the transient class unit tests for wiserep
