### 2026-01-23 427eba1
add a new example for accessing wiserep spectra
