### 2026-01-23 4259496
fix: raising the query wiserep exception to keep track of the full traceback
