### 2026-01-23 002d951
skip taahe DataFinder.query_wisrep metest since it queries wiserep too fast back to back and then fails
