### 2026-01-23 1b1dbe2
fix: minor issue with pytest decorator
