### 2026-01-26 6967e2d
adjust wiserep tests further
