### 2026-01-26 efdc614
fix a bug in a now-deprecated functionality of astropy units
