### 2026-01-26 b74436e
fix: a NaN bug in _merge_arbitrary
