### 2026-01-26 52ad10f
fix: edit pandas version requirement for wiserep_api
