+++ tests/test_data_finder.py
@@ -270,9 +270,6 @@ def test_query_heasarc():
     assert len(res2) >= 2, "Missing some HEASARC data"
 
 
-@pytest.mark.skip(
-    reason="this test crashes wiserep because too many queries back to back"
-)
 def test_query_wiserep():
     """
     Test querying WISeREP for spectra
