+++ tests/test_transient.py
@@ -322,9 +322,6 @@ def test_clean_photometry():
     assert np.isclose(phot_non_default.converted_flux.iloc[0], 2.69e3), msg
 
 
-@pytest.mark.skip(
-    "WISeREP queries on GitHub seem to consistently crash, but not locally!"
-)
 def test_get_wiserep_spec():
     test_t = Transient(
         {
