### 2026-01-27 33ceb2e
v1.1.0
Automatically generated by python-semantic-release [skip actions]
