### 2026-02-02 #9516 @noahfranz13
I just got the proofs for the AAS article.

@warrickball it looks like the DOI will be 10.3847/1538-4357/ae346e. I'm not sure if we need to wait for the AAS article to actually be published though since this DOI is not active?

@augustfly The AAS journal proofs do not have the link the the JOSS DOI (it does not have the blurb after the abstract saying "Software reviewed by the Journal of Open Source Software"). Is there something else I/we need to do to get it on there?
