### 2026-02-02 #9516 @warrickball
> @warrickball it looks like the DOI will be 10.3847/1538-4357/ae346e. I'm not sure if we need to wait for the AAS article to actually be published though since this DOI is not active?

Could you go ahead and add the DOI to the `aas-doi` field in the JOSS paper's metadata? I can then do a fresh test run.

If I understand correctly, I think it's tidier if we publish the JOSS article because it then fixes our DOI for inclusion in the AAS Journals article. If we need to change anything about the JOSS paper, we can re-accept anyway.
