### 2026-02-02 #9516 @augustfly
@noahfranz13 Ugh, I put my notes about the JOSS article in the wrong place in our system (on your manuscript "R1" and not on the final "R2" version), and it was completely overlooked. This can be fixed. I'll send an email to the ApJ production office right now.  Thanks for drawing my attention to it.

On your ApJ proof you'll want to try to figure a place to cite the JOSS article. That would be in addition to the banner, which I'll get added. In the right workflow, Greg would have edited the manuscript preprint with you on this at the data proof to add the JOSS citation, but he never saw my comments either. Apologies.
