### 2026-02-02 fd3df58
Update aas-doi with the value from the AAS journal proofs
