### 2026-02-03 #9516 @noahfranz13
> Could you go ahead and add the DOI to the aas-doi field in the JOSS paper's metadata? I can then do a fresh test run.

@warrickball this should be done now!

> On your ApJ proof you'll want to try to figure a place to cite the JOSS article.

@augustfly I added a citation to the JOSS paper to the proof.

 @warrickball I cited this JOSS doi: 10.21105/joss.09516 from the PDF generated above, and I just wanted to confirm that was correct still?
