### 2026-02-03 #9516 @warrickball
Sorry, you're absolutely right that the JOSS DOI is correct. It's the volume and issue that are set by year and month.
