### 2026-02-03 #9516 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
