### 2026-02-03 #9516 @warrickball
Congratulations @noahfranz13, your paper has been published in JOSS! Once the AAS Journals paper is up, you should check that the DOI in our PDF resolves and comment here to let me know if there's any issue. (I'll try to remember to have a look myself, too.) I'll close the issue now to mark our process as complete but you'll still be able to comment.

Though it might already feel like a distant memory, thanks once again to @erusseil, @apace7 this submission! JOSS simply wouldn't be possible without its community of volunteers.
